# [NUI] Do not call default creator when we set value