"ALL" default BindableProperty.PropertyChanged events what NUI using, dont use 'oldValue'. But due to that usecase, we always call DefaultValueCreator when we try to set value.

To make NUI platform faster, let we ignore that property getter when we only use setter.

For XMAL case, support for backward stabilize, let we don't use this feature.

Performance result for app `Tizen.NUI.PerformanceTest`

(dali version : 2.3.5. Release, Execute on WSL server based on Ubuntu22.04, i7 dual core)

|Creation Type|Before PR (ms)|After PR (ms)|Reduce Rate (1 - after / before)| 
|-------------|-------------|-------------|-------------| 
| 40 COLOR | 2.6657574074074066 | 2.19945 | 17.492492232 % | 
| 40 IMAGE | 1.9131259259259266 | 1.56895 | 17.99023897 % | 
| 40 TEXT | 7.013750000000001 | 6.103418518518519 | 12.979240513 % |
| 40 ROUNDED COLOR | 3.342475925925925 | 2.8528203703703694 | 14.649486381 % |
 | 40 BORDER COLOR | 4.039824074074074 | 3.8768833333333332 | 4.033362289 % |
 | 40 ROUNDED BORDER COLOR | 4.641790740740741 | 4.230518518518518 | 8.860206011 % | 
| 40 BLUR COLOR | 2.7265814814814813 | 2.329972222222222 | 14.546026295 % | 
| 40 ROUNDED BLUR COLOR | 2.817055555555555 | 2.56367962962963 | 8.994353179 % |