@hinohie 2024-01-06
Hello @elishateng @AdunFang.

I'd like to check whether this way to fix is looks okay.

To support it correctly, we'd better add some flag something like `IgnoreDefaultValueCreatorIfSetValueFirstTime` (Default is false) at `BindableProperty` class, and add that flags for all `BindableProperty` what NUI use default. But this class is relative with `Tizen.NUI.XamlBuild` repository, so I think add new flag / property is quite dangerous.

Could you check whether this PR has no problem even if we use XAML cases?
+ Could you check that add the property at `BindableProperty`?