@elishateng 2024-01-08
> Hello @elishateng @AdunFang.
> 
> I'd like to check whether this way to fix is looks okay.
> 
> To support it correctly, we'd better add some flag something like `IgnoreDefaultValueCreatorIfSetValueFirstTime` (Default is false) at `BindableProperty` class, and add that flags for all `BindableProperty` what NUI use default. But this class is relative with `Tizen.NUI.XamlBuild` repository, so I think add new flag / property is quite dangerous.
> 
> Could you check whether this PR has no problem even if we use XAML cases?
> 
> * Could you check that add the property at `BindableProperty`?


Hi Mr. Hong,
Here is a sample of Text BindableProperty of TextLabel:

```.cs
        public static readonly BindableProperty TextProperty = BindableProperty.Create(nameof(Text), typeof(string), typeof(TextLabel), string.Empty, propertyChanged: (BindableProperty.BindingPropertyChangedDelegate)((bindable, oldValue, newValue) =>
        {
            var textLabel = (TextLabel)bindable;

            if (newValue is Selector<string> selector)
            {
                textLabel.TextSelector = selector;
            }
            else
            {
                textLabel.selectorData?.Text?.Reset(textLabel);
                textLabel.SetText((string)newValue);
            }
        }),
        defaultValueCreator: (BindableProperty.CreateDefaultValueDelegate)((bindable) =>
        {
            var textLabel = (TextLabel)bindable;

            // Do not try to get string if we know that previous text was empty.
            return textLabel.textIsEmpty ? "" : Object.InternalGetPropertyString(textLabel.SwigCPtr, TextLabel.Property.TEXT);
        }));
```


This is the result of my talk with Mr. Fang, let’s take TextProperty as an example: 
 
1. Initially the property.DefaultValue of Text is string.Empty,  we call SetValue(TextProperty), then the property.DefaultValue assigns to oldvalue, it may not the default value in Dali side.  //new TextLabel("Test")

2. After calling Setvalue again, if the property changed from dali side, like we do animation for the Text property, then the property value may different from previous value of SetValue.

3. In xaml binding, the property value also can changed from dali. It's same with above 2th item.

4. Above all, we recommand use DefaultValueCreator to get the real value from dali side.