@hinohie 2024-01-09
@elishateng 
This patch only consider "SetValue" cases. (We will use DefaultCreator for GetValue cases)

Let we think about only `SetValue` cases.
`BindableObject` Try to get value from context, by `property.ValueGetter(this)` at `BindablePropertyContext GetOrCreateContext` line 903 or 907. And then, `context.Value` become sync with DALi side.

But... If we call `InternalSetValue` when `IsBound = false`. We set `context.Value` and send this value as `BindableProperty.PropertyChanged`'s oldValue.

The review point of this patch is, we can ignore that `oldValue` since NUI don't use it now.
(For xaml binding case, IsBound might be true so It will go to `SetValueCore`, where I make `ignoreDefaultValueCreator : false` now.)

This is the sample code for example,

```[code=c#]
            control.Position = new Position(0.0f, 0.0f);
            var position = control.Position;
            Tizen.Log.Error("NUI", $"Positoin : {position.X} {position.Y}\n");
            position = control.Position;
            Tizen.Log.Error("NUI", $"Positoin : {position.X} {position.Y}\n");
            control.Position = new Position(2.0f, 2.0f);
            position = control.Position;
            Tizen.Log.Error("NUI", $"Positoin : {position.X} {position.Y}\n");
```

Before this patch :
```
Before

ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1009) > SetPosition
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ControlSample.cs: Activate(40) > Positoin : 0 0
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ControlSample.cs: Activate(42) > Positoin : 0 0
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1009) > SetPosition
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ControlSample.cs: Activate(45) > Positoin : 2 2
```

After this patch :
```
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1009) > SetPosition
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ControlSample.cs: Activate(40) > Positoin : 0 0
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ControlSample.cs: Activate(42) > Positoin : 0 0
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1009) > SetPosition
ERROR: NUI: ViewBindableProperty.cs: PositionProperty(1018) > GetPosition
ERROR: NUI: ControlSample.cs: Activate(45) > Positoin : 2 2
```



Please check this PR only for `SetValue` cases.